Refacto to break apart 1000 line mod file, renamed `vergen-fmt` to 'vergen-pretty'
