+++ .github/workflows/main.yml
@@ -53,7 +53,7 @@ jobs:
           cargo clippy -p vergen --all-targets --features build,cargo,git,gitcl,rustc,si -- -Dwarnings
           cargo clippy -p vergen --all-targets --features build,cargo,git,git2,rustc,si -- -Dwarnings
           cargo clippy -p vergen --all-targets --features build,cargo,git,gix,rustc,si -- -Dwarnings
-          cargo clippy -p vergen-fmt --all-targets --all-features -- -Dwarnings
+          cargo clippy -p vergen-pretty --all-targets --all-features -- -Dwarnings
   lints-macos:
     name: 🕳️ Lints MacOS 🕳️
     needs: rustfmt
@@ -81,7 +81,7 @@ jobs:
           cargo clippy -p vergen --all-targets --features build,cargo,git,gitcl,rustc,si -- -Dwarnings
           cargo clippy -p vergen --all-targets --features build,cargo,git,git2,rustc,si -- -Dwarnings
           cargo clippy -p vergen --all-targets --features build,cargo,git,gix,rustc,si -- -Dwarnings
-          cargo clippy -p vergen-fmt --all-targets --all-features -- -Dwarnings
+          cargo clippy -p vergen-pretty --all-targets --all-features -- -Dwarnings
   lints-windows:
     name: 🕳️ Lints Windows 🕳️
     needs: rustfmt
@@ -109,7 +109,7 @@ jobs:
           cargo clippy -p vergen --all-targets --features build,cargo,git,gitcl,rustc,si -- -Dwarnings
           cargo clippy -p vergen --all-targets --features build,cargo,git,git2,rustc,si -- -Dwarnings
           cargo clippy -p vergen --all-targets --features build,cargo,git,gix,rustc,si -- -Dwarnings
-          cargo clippy -p vergen-fmt --all-targets --all-features -- -Dwarnings
+          cargo clippy -p vergen-pretty --all-targets --all-features -- -Dwarnings
   test-linux:
     name: 🧪 Linux 🧪
     needs: lints-linux
@@ -158,18 +158,18 @@ jobs:
           cargo test -p vergen -F build,cargo,git,gitcl,rustc,si
           cargo test -p vergen -F build,cargo,git,git2,rustc,si
           cargo test -p vergen -F build,cargo,git,gix,rustc,si
-          cargo test -p vergen-fmt -F __vergen_test
-          cargo test -p vergen-fmt -F __vergen_test,color
-          cargo test -p vergen-fmt -F __vergen_test,trace
-          cargo test -p vergen-fmt -F __vergen_test,color,trace
+          cargo test -p vergen-pretty -F __vergen_test
+          cargo test -p vergen-pretty -F __vergen_test,color
+          cargo test -p vergen-pretty -F __vergen_test,trace
+          cargo test -p vergen-pretty -F __vergen_test,color,trace
       - name: 🙊 Code Coverage (vergen) 🙊
-        run: cargo tarpaulin --config vergen/tarpaulin.toml --exclude-files vergen/src/lib.rs --exclude-files vergen-fmt/* --out Xml --output-dir vergen
-      - name: 🙊 Code Coverage (vergen-fmt) 🙊
-        run: cargo tarpaulin --config vergen-fmt/tarpaulin.toml --exclude-files vergen-fmt/src/lib.rs --exclude-files vergen/* --out Xml --output-dir vergen-fmt
+        run: cargo tarpaulin --config vergen/tarpaulin.toml --exclude-files vergen/src/lib.rs --exclude-files vergen-pretty/* --out Xml --output-dir vergen
+      - name: 🙊 Code Coverage (vergen-pretty) 🙊
+        run: cargo tarpaulin --config vergen-pretty/tarpaulin.toml --exclude-files vergen-pretty/src/lib.rs --exclude-files vergen/* --out Xml --output-dir vergen-pretty
       - name: 🙊 Upload Code Coverage 🙊
         uses: codecov/codecov-action@v3
         with:
-          files: ./vergen/cobertura.xml,./vergen-fmt/cobertura.xml
+          files: ./vergen/cobertura.xml,./vergen-pretty/cobertura.xml
   test-macos:
     name: 🧪 MacOS 🧪
     needs: lints-macos
@@ -217,10 +217,10 @@ jobs:
           cargo test -p vergen -F build,cargo,git,gitcl,rustc,si
           cargo test -p vergen -F build,cargo,git,git2,rustc,si
           cargo test -p vergen -F build,cargo,git,gix,rustc,si
-          cargo test -p vergen-fmt -F __vergen_test
-          cargo test -p vergen-fmt -F __vergen_test,color
-          cargo test -p vergen-fmt -F __vergen_test,trace
-          cargo test -p vergen-fmt -F __vergen_test,color,trace
+          cargo test -p vergen-pretty -F __vergen_test
+          cargo test -p vergen-pretty -F __vergen_test,color
+          cargo test -p vergen-pretty -F __vergen_test,trace
+          cargo test -p vergen-pretty -F __vergen_test,color,trace
   test-windows:
     name: 🧪 Windows 🧪
     needs: lints-windows
@@ -280,7 +280,7 @@ jobs:
           cargo test -p vergen -F build,cargo,git,gitcl,rustc,si
           cargo test -p vergen -F build,cargo,git,git2,rustc,si
           cargo test -p vergen -F build,cargo,git,gix,rustc,si
-          cargo test -p vergen-fmt -F __vergen_test
-          cargo test -p vergen-fmt -F __vergen_test,color
-          cargo test -p vergen-fmt -F __vergen_test,trace
-          cargo test -p vergen-fmt -F __vergen_test,color,trace
+          cargo test -p vergen-pretty -F __vergen_test
+          cargo test -p vergen-pretty -F __vergen_test,color
+          cargo test -p vergen-pretty -F __vergen_test,trace
+          cargo test -p vergen-pretty -F __vergen_test,color,trace
